requirements.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile --output-file requirements.txt requirements.in
#
appnope==0.1.0            # via ipython
attrs==18.2.0
backcall==0.1.0           # via ipython
bleach==3.0.2             # via nbconvert
boto3==1.9.71
botocore==1.12.71         # via boto3, s3transfer
certifi==2018.11.29       # via requests
chardet==3.0.4            # via requests
cycler==0.10.0            # via matplotlib
decorator==4.3.0          # via ipython, traitlets
defusedxml==0.5.0         # via nbconvert
django-environ==0.4.5
django==2.1.4
docutils==0.14            # via botocore
dpath==1.4.2
entrypoints==0.2.3        # via nbconvert
idna==2.8                 # via requests
ipykernel==5.1.0          # via ipywidgets, jupyter, jupyter-console, notebook, qtconsole
ipython-genutils==0.2.0   # via nbformat, notebook, qtconsole, traitlets
ipython==7.2.0
ipywidgets==7.4.2         # via jupyter
jedi==0.13.2              # via ipython
jinja2==2.10              # via nbconvert, notebook
jmespath==0.9.3           # via boto3, botocore
jsonschema==2.6.0         # via nbformat
jupyter-client==5.2.4     # via ipykernel, jupyter-console, notebook, qtconsole
jupyter-console==6.0.0    # via jupyter
jupyter-core==4.4.0       # via jupyter-client, nbconvert, nbformat, notebook, qtconsole
jupyter==1.0.0
kiwisolver==1.0.1         # via matplotlib
markupsafe==1.1.0         # via jinja2
matplotlib==3.0.2
mistune==0.8.4            # via nbconvert
nbconvert==5.4.0          # via jupyter, notebook
nbformat==4.4.0           # via ipywidgets, nbconvert, notebook
notebook==5.7.4           # via jupyter, widgetsnbextension
numpy==1.15.4             # via matplotlib, pandas
pandas==0.23.4
pandocfilters==1.4.2      # via nbconvert
parso==0.3.1              # via jedi
pexpect==4.6.0            # via ipython
pickleshare==0.7.5        # via ipython
prometheus-client==0.5.0  # via notebook
prompt-toolkit==2.0.7     # via ipython, jupyter-console
ptyprocess==0.6.0         # via pexpect, terminado
pygments==2.3.1           # via ipython, jupyter-console, nbconvert, qtconsole
pyparsing==2.3.0          # via matplotlib
python-dateutil==2.7.5    # via botocore, jupyter-client, matplotlib, pandas
pytz==2018.7              # via django, pandas
pyyaml==3.13
pyzmq==17.1.2             # via jupyter-client, notebook
qtconsole==4.4.3          # via jupyter
requests==2.21.0
s3transfer==0.1.13        # via boto3
send2trash==1.5.0         # via notebook
six==1.12.0               # via bleach, cycler, prompt-toolkit, python-dateutil, traitlets
terminado==0.8.1          # via notebook
testpath==0.4.2           # via nbconvert
tornado==5.1.1            # via ipykernel, jupyter-client, notebook, terminado
tqdm==4.28.1
traitlets==4.3.2          # via ipykernel, ipython, ipywidgets, jupyter-client, jupyter-core, nbconvert, nbformat, notebook, qtconsole
urllib3==1.24.1           # via botocore, requests
wcwidth==0.1.7            # via prompt-toolkit
webencodings==0.5.1       # via bleach
widgetsnbextension==3.4.2  # via ipywidgets
